backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-channels==0.7.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.1
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.4
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
zope.interface==8.1.1
